# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shin SuperApp",
    "version": "2.1.0"
  },
  "paths": {
...
    "/api/catalog/categories": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Categories",
        "description": "List all app categories, merged from DB and app registry.",
        "operationId": "list_categories_api_catalog_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AppCategoryRead"
                  },
                  "type": "array",
                  "title": "Response List Categories Api Catalog Categories Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Create Category",
        "description": "Create a new app category. Requires admin.",
        "operationId": "create_category_api_catalog_categories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppCategoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/catalog/categories/{category_id}": {
      "patch": {
        "tags": [
          "catalog"
        ],
        "summary": "Update Category",
        "description": "Update an app category.",
        "operationId": "update_category_api_catalog_categories__category_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppCategoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "catalog"
        ],
        "summary": "Delete Category",
        "description": "Delete an app category.",
        "operationId": "delete_category_api_catalog_categories__category_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Category Api Catalog Categories  Category Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalog": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Catalog",
        "description": "List all available apps, marking installed apps for authenticated users.",
        "operationId": "list_catalog_api_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AppCatalogEntry"
                  },
                  "type": "array",
                  "title": "Response List Catalog Api Catalog Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/catalog/install": {
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Install App",
        "description": "Install an app for the current user.",
        "operationId": "install_app_api_catalog_install_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppInstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Install App Api Catalog Install Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/catalog/uninstall": {
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Uninstall App",
        "description": "Uninstall an app for the current user (soft delete).",
        "operationId": "uninstall_app_api_catalog_uninstall_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUninstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Uninstall App Api Catalog Uninstall Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/catalog/preferences": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get All Preferences",
        "description": "Get all widget preferences for the authenticated user across all apps.",
        "operationId": "get_all_preferences_api_catalog_preferences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WidgetPreferenceSchema"
                  },
                  "type": "array",
                  "title": "Response Get All Preferences Api Catalog Preferences Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/catalog/preferences/{app_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Preferences",
        "description": "Get all widget preferences for the authenticated user in a given app.",
        "operationId": "get_preferences_api_catalog_preferences__app_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WidgetPreferenceSchema"
                  },
                  "title": "Response Get Preferences Api Catalog Preferences  App Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "catalog"
        ],
        "summary": "Update Preferences",
        "description": "Batch-update widget preferences using shared utility.",
        "operationId": "update_preferences_api_catalog_preferences__app_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PreferenceUpdate"
                },
                "title": "Updates"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WidgetPreferenceSchema"
                  },
                  "title": "Response Update Preferences Api Catalog Preferences  App Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppCatalogEntry": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "is_installed": {
            "type": "boolean",
            "title": "Is Installed",
            "default": false
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "screenshots": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Screenshots",
            "default": []
          },
          "widgets": {
            "items": {
              "$ref": "#/components/schemas/WidgetManifestSchema"
            },
            "type": "array",
            "title": "Widgets"
          },
          "requires_tier": {
            "type": "string",
            "enum": [
              "free",
              "paid"
            ],
            "title": "Requires Tier",
            "default": "free"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "icon",
          "color",
          "category",
          "version",
          "author"
        ],
        "title": "AppCatalogEntry",
        "description": "Summary entry shown in the app store."
      },
      "AppCategoryRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "order": {
            "type": "integer",
            "title": "Order"
          },
          "auto_discovered": {
            "type": "boolean",
            "title": "Auto Discovered",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "icon",
          "color",
          "order"
        ],
        "title": "AppCategoryRead",
        "description": "App catalog category entry shown in the store UI."
      },
      "AppInstallRequest": {
        "properties": {
          "app_id": {
            "type": "string",
            "title": "App Id"
          }
        },
        "type": "object",
        "required": [
          "app_id"
        ],
        "title": "AppInstallRequest"
      },
...
      "AppUninstallRequest": {
        "properties": {
          "app_id": {
            "type": "string",
            "title": "App Id"
          }
        },
        "type": "object",
        "required": [
          "app_id"
        ],
        "title": "AppUninstallRequest"
      },
...
      "ConfigFieldSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Field key, camelCase"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "number",
              "select",
              "multi-select",
              "date",
              "boolean"
            ],
            "title": "Type"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "default": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/SelectOption"
            },
            "type": "array",
            "title": "Options"
          },
          "options_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options Source",
            "description": "Widget resolver ID for dynamic options, e.g. 'finance.wallets'"
          },
          "placeholder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placeholder"
          },
          "min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min"
          },
          "max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max"
          },
          "step": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          }
        },
        "type": "object",
        "required": [
          "name",
          "label",
          "type"
        ],
        "title": "ConfigFieldSchema",
        "description": "A configuration field for a widget."
      },
      "CreateCategoryRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name"
          },
          "icon": {
            "type": "string",
            "maxLength": 30,
            "title": "Icon",
            "default": "Folder"
          },
          "color": {
            "type": "string",
            "maxLength": 50,
            "title": "Color",
            "default": "oklch(0.65 0.21 280)"
          },
          "order": {
            "type": "integer",
            "title": "Order",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateCategoryRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreferenceUpdate": {
        "properties": {
          "widget_id": {
            "type": "string",
            "title": "Widget Id"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "size_w": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12.0,
                "minimum": 2.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Size W"
          },
          "size_h": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 6.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Size H"
          }
        },
        "type": "object",
        "required": [
          "widget_id"
        ],
        "title": "PreferenceUpdate",
        "description": "Update payload for a single widget preference."
      },
...
      "UpdateCategoryRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order"
          }
        },
        "type": "object",
        "title": "UpdateCategoryRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WidgetManifestSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique widget ID, format '{app_id}.{widget_name}', e.g. 'finance.total-balance'"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "icon": {
            "type": "string",
            "title": "Icon",
            "description": "Lucide icon name"
          },
          "size": {
            "type": "string",
            "enum": [
              "compact",
              "standard",
              "wide",
              "tall",
              "full"
            ],
            "title": "Size",
            "default": "standard"
          },
          "config_fields": {
            "items": {
              "$ref": "#/components/schemas/ConfigFieldSchema"
            },
            "type": "array",
            "title": "Config Fields"
          },
          "requires_auth": {
            "type": "boolean",
            "title": "Requires Auth",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "icon"
        ],
        "title": "WidgetManifestSchema",
        "description": "A single widget definition \u2014 referenced in AppManifestSchema.widgets."
      },
      "WidgetPreferenceSchema": {
        "properties": {
          "_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "widget_id": {
            "type": "string",
            "title": "Widget Id"
          },
          "app_id": {
            "type": "string",
            "title": "App Id"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "size_w": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12.0,
                "minimum": 2.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Size W"
          },
          "size_h": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 6.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Size H"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "widget_id",
          "app_id"
        ],
        "title": "WidgetPreferenceSchema",
        "description": "User-specific widget state stored in DB."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}